Split quota plan CLI smoke assertions
